some refactoring of the BackwardEuler solver
