+++ integration/BackwardEuler/Make.package
@@ -1,1 +1,4 @@
 CEXE_headers += actual_integrator.H
+CEXE_headers += be_integrator.H
+CEXE_headers += be_type.H
+CEXE_headers += be_type_strang.H
